wrapper refactor
