+++ environments/general_wrappers.py
@@ -5,12 +5,12 @@
 from ppo_and_friends.utils.stats import RunningMeanStd
 import numpy as np
 import pickle
+import sys
 import os
 from ppo_and_friends.utils.mpi_utils import rank_print
 from ppo_and_friends.utils.misc import need_action_squeeze
 from collections.abc import Iterable
 from gym.spaces import Box, Discrete, MultiDiscrete, MultiBinary
-from gym.spaces.tuple import Tuple as gym_tuple
 from mpi4py import MPI
 
 comm      = MPI.COMM_WORLD
@@ -18,7 +18,7 @@ rank      = comm.Get_rank()
 num_procs = comm.Get_size()
 
 
-class IdentityWrapper(object):
+class IdentityWrapper():
     """
         A wrapper that acts exactly like the original environment but also
         has a few extra bells and whistles.
@@ -35,6 +35,7 @@ class IdentityWrapper(object):
                 env           The environment to wrap.
                 status_dict   The dictionary containing training stats.
         """
+        super(IdentityWrapper, self).__init__(**kw_args)
 
         self.env               = env
         self.test_mode         = test_mode
@@ -115,6 +116,43 @@ class IdentityWrapper(object):
         obs = self.env.reset()
         return obs
 
+    def _env_soft_reset(self):
+        """
+            Check if our enviornment can perform a soft reset. If so, do it.
+            If not, perform a standard reset.
+
+            Returns:
+                If our environment has a soft_reset method, we return the result
+                of calling env.soft_reset. Otherwise, we return our obs_cache.
+        """
+        #
+        # Tricky Business:
+        # Perform a recursive call on all wrapped environments to check
+        # for an ability to soft reset. Here's an example of how this works
+        # in practice:
+        #
+        #   Imagine we wrap our environment like so:
+        #   env = MultiAgentWrapper(ObservationNormalizer(ObservationClipper(env)))
+        #
+        #   When we call 'env.soft_reset()', we are calling
+        #   ObservationClipper.soft_reset(). In this case, though, the clipper
+        #   does not have direct access to the global observations from the
+        #   MultiAgentWrapper. So, we recursively travel back to the base
+        #   wrapper, and return its soft_reset() results. Those results will
+        #   then travel back up through the normalizer and clipper, being
+        #   normalized and clipped, before being returned. It's basically
+        #   identical to our "step" and "reset" calls, except each wrapper
+        #   may have its own cache. The takeaway is that we ignore all caches
+        #   other than the bottom level cache, which is transformed on its
+        #   way out of the depths of recursion.
+        #
+        soft_reset = getattr(self.env, "soft_reset", None)
+
+        if callable(soft_reset):
+            return soft_reset()
+
+        return self.obs_cache
+
     def soft_reset(self):
         """
             Perform a "soft reset". This results in only performing the reset
@@ -126,19 +164,9 @@ class IdentityWrapper(object):
                 An observation.
         """
         if self.need_hard_reset or type(self.obs_cache) == type(None):
-
-            #
-            # Perform a recursive call on all wrapped environments to check
-            # for an ability to soft reset.
-            #
-            soft_reset = getattr(self.env, "soft_reset", None)
-
-            if callable(soft_reset):
-                return soft_reset()
-
             return self.reset()
 
-        return self.obs_cache
+        return self._env_soft_reset()
 
     def render(self, **kw_args):
         """
@@ -235,6 +263,7 @@ class IdentityWrapper(object):
             return batch_size_getter()
         else:
             return 1
+
 
 class VectorizedEnv(IdentityWrapper, Iterable):
     """
@@ -940,6 +969,7 @@ class MultiAgentWrapper(IdentityWrapper):
         """
         obs = self.env.reset()
         obs = self._refine_obs(obs)
+
         global_state = self.get_feature_pruned_global_state(obs)
 
         return obs, global_state
@@ -958,6 +988,7 @@ class MultiAgentWrapper(IdentityWrapper):
             return self.reset()
 
         global_state = self.get_feature_pruned_global_state(self.obs_cache)
+
         return self.obs_cache, global_state
 
     def get_batch_size(self):
@@ -968,711 +999,6 @@ class MultiAgentWrapper(IdentityWrapper):
                 Return our batch size.
         """
         return self.num_agents
-
-    def supports_batched_environments(self):
-        """
-            Determine whether or not our wrapped environment supports
-            batched environments.
-
-            Return:
-                True.
-        """
-        return True
-
-
-class ObservationNormalizer(IdentityWrapper):
-    """
-        A wrapper for normalizing observations. This normalization
-        method uses running statistics.
-        NOTE: this uses implentations very similar to some of Stable
-        Baslines' VecNormalize.
-    """
-
-    def __init__(self,
-                 env,
-                 update_stats = True,
-                 epsilon      = 1e-8,
-                 **kw_args):
-        """
-            Initialize the wrapper.
-
-            Arguments:
-                env            The environment to wrap.
-                update_stats   Should we update the statistics? We typically
-                               have this enabled until testing.
-                epsilon        A very small number to help avoid dividing by 0.
-        """
-
-        super(ObservationNormalizer, self).__init__(
-            env,
-            **kw_args)
-
-        self.running_stats = RunningMeanStd(
-            shape = self.env.observation_space.shape)
-
-        self.update_stats  = update_stats
-        self.epsilon       = epsilon
-
-    def step(self, action):
-        """
-            Take a single step in the environment using the given
-            action, update the running stats, and normalize the
-            resulting observation.
-
-            Arguments:
-                action    The action to take.
-
-            Returns:
-                The resulting observation, reward, done, and info tuple.
-        """
-        obs, reward, done, info = self.env.step(action)
-
-        if self.update_stats:
-            self.running_stats.update(obs)
-
-        if type(obs) == np.ndarray:
-            obs = self.normalize(obs.copy())
-        else:
-            obs = self.normalize(obs)
-
-        return obs, reward, done, info
-
-    def reset(self):
-        """
-            Reset the environment, and update the running stats.
-
-            Returns:
-                The resulting observation.
-        """
-        obs = self.env.reset()
-
-        #
-        # In our multi-agent case, the global state is returned along
-        # with the local observation.
-        #
-        is_multi_agent = False
-        if type(obs) == tuple  or type(obs) == list:
-            is_multi_agent = True
-            obs, global_state = obs
-
-        if self.update_stats:
-            self.running_stats.update(obs)
-
-        if is_multi_agent:
-            return (obs, global_state)
-        return obs
-
-    def normalize(self, obs):
-        """
-            Normalize an observation using our running stats.
-
-            Arguments:
-                obs    The observation to normalize.
-
-            Returns:
-                The normalized observation.
-        """
-        obs = (obs - self.running_stats.mean) / \
-            np.sqrt(self.running_stats.variance + self.epsilon)
-        return obs
-
-    def save_info(self, path):
-        """
-            Save out our running stats, and check if our wrapped
-            environment needs to perform any more info saves.
-
-            Arguments:
-                path    The path to save to.
-        """
-        if self.test_mode:
-            return
-
-        file_name = "RunningObsStats_{}.pkl".format(rank)
-        out_file  = os.path.join(path, file_name)
-
-        with open(out_file, "wb") as fh:
-            pickle.dump(self.running_stats, fh)
-
-        self._check_env_save(path)
-
-    def load_info(self, path):
-        """
-            Load our running stats and check to see if our wrapped
-            environment needs to load anything.
-
-            Arguments:
-                path    The path to load from.
-        """
-        if self.test_mode:
-            file_name = "RunningObsStats_0.pkl"
-        else:
-            file_name = "RunningObsStats_{}.pkl".format(rank)
-
-        in_file = os.path.join(path, file_name)
-
-        #
-        # There are cases where we initially train using X ranks, and we
-        # later want to continue training using (X+k) ranks. In these cases,
-        # let's copy rank 0's info to all ranks > X.
-        #
-        if not os.path.exists(in_file):
-            file_name = "RunningObsStats_0.pkl"
-            in_file   = os.path.join(path, file_name)
-
-        with open(in_file, "rb") as fh:
-            self.running_stats = pickle.load(fh)
-
-        self._check_env_load(path)
-
-
-class RewardNormalizer(IdentityWrapper):
-    """
-        This wrapper uses running statistics to normalize rewards.
-        NOTE: some of this logic comes from Stable Baseline's
-        VecNormalize.
-    """
-
-    def __init__(self,
-                 env,
-                 update_stats = True,
-                 epsilon      = 1e-8,
-                 gamma        = 0.99,
-                 **kw_args):
-        """
-            Initialize the wrapper.
-
-            Arguments:
-                env            The environment to wrap.
-                update_stats   Whether or not to update our running stats. This
-                               is typically set to true until testing.
-                epsilon        A very small value to help avoid 0 divisions.
-                gamma          A discount factor for our running stats.
-        """
-        super(RewardNormalizer, self).__init__(
-            env,
-            **kw_args)
-
-        self.running_stats = RunningMeanStd(shape=())
-        self.update_stats  = update_stats
-        self.epsilon       = epsilon
-        self.gamma         = gamma
-
-        #
-        # We might be wrapping an environment that supports batches.
-        # if so, we need to be able to correlate a running reward with
-        # each environment instance.
-        #
-        if self.test_mode and self.get_num_agents() == 1:
-            self.batch_size = 1
-        else:
-            self.batch_size = self.get_batch_size()
-
-        self.running_reward = np.zeros(self.batch_size)
-
-    def step(self, action):
-        """
-            Take a single step in the environment using the given
-            action, update the running stats, and normalize the
-            resulting reward.
-
-            Arguments:
-                action    The action to take.
-
-            Returns:
-                The resulting observation, reward, done, and info tuple.
-        """
-
-        obs, reward, done, info = self.env.step(action)
-
-        for env_idx in range(self.batch_size):
-            if self.update_stats:
-                self.running_reward[env_idx] = \
-                    self.running_reward[env_idx] * self.gamma + reward[env_idx]
-
-                self.running_stats.update(self.running_reward)
-
-        if self.batch_size > 1:
-            where_done = np.where(done)[0]
-            self.running_reward[where_done] = 0.0
-        elif done:
-            self.running_reward[0] = 0.0
-
-        if type(reward) == np.ndarray:
-            batch_size = reward.shape[0]
-
-            for r_idx in range(batch_size):
-                if "natural reward" not in info[r_idx]:
-                    info[r_idx]["natural reward"] = reward[r_idx]
-
-            reward = self.normalize(reward.copy())
-
-        else:
-            if "natural reward" not in info:
-                info["natural reward"] = reward
-
-            reward = self.normalize(reward)
-
-        return obs, reward, done, info
-
-    def normalize(self, reward):
-        """
-            Normalize our reward using Stable Baseline's approach.
-
-            Arguments:
-                reward    The reward to normalize.
-
-            Returns:
-                The normalized reward.
-        """
-        reward /= np.sqrt(self.running_stats.variance + self.epsilon)
-        return reward
-
-    def save_info(self, path):
-        """
-            Save out our running stats, and check if our wrapped
-            environment needs to perform any more info saves.
-
-            Arguments:
-                path    The path to save to.
-        """
-        if self.test_mode:
-            return
-
-        file_name = "RunningRewardsStats_{}.pkl".format(rank)
-        out_file  = os.path.join(path, file_name)
-
-        with open(out_file, "wb") as fh:
-            pickle.dump(self.running_stats, fh)
-
-        self._check_env_save(path)
-
-    def load_info(self, path):
-        """
-            Load our running stats and check to see if our wrapped
-            environment needs to load anything.
-
-            Arguments:
-                path    The path to load from.
-        """
-        if self.test_mode:
-            file_name = "RunningRewardsStats_0.pkl"
-        else:
-            file_name = "RunningRewardsStats_{}.pkl".format(rank)
-
-        in_file = os.path.join(path, file_name)
-
-        #
-        # There are cases where we initially train using X ranks, and we
-        # later want to continue training using (X+k) ranks. In these cases,
-        # let's copy rank 0's info to all ranks > X.
-        #
-        if not os.path.exists(in_file):
-            file_name = "RunningRewardsStats_0.pkl"
-            in_file   = os.path.join(path, file_name)
-
-        with open(in_file, "rb") as fh:
-            self.running_stats = pickle.load(fh)
-
-        self._check_env_load(path)
-
-
-class GenericClipper(IdentityWrapper):
-    """
-        A wrapper for clipping rewards.
-    """
-
-    def __init__(self,
-                 env,
-                 status_dict = {},
-                 clip_range  = (-10., 10.),
-                 **kw_args):
-        """
-            Initialize the wrapper.
-
-            Arguments:
-                env         The environment to wrap.
-                status_dict The training status dictionary. This is used when
-                            our clip range contains callables.
-                clip_range  The range to clip our rewards into. This can be
-                            either a real number or a class that inherits from
-                            IterationMapper.
-        """
-
-        super(GenericClipper, self).__init__(
-            env,
-            **kw_args)
-
-        min_callable = None
-        max_callable = None
-        self.status_dict = status_dict
-
-        if callable(clip_range[0]):
-            min_callable = clip_range[0]
-        else:
-            min_callable = lambda *args, **kwargs : clip_range[0]
-
-        if callable(clip_range[1]):
-            max_callable = clip_range[1]
-        else:
-            max_callable = lambda *args, **kwargs : clip_range[1]
-
-        self.clip_range = (min_callable, max_callable)
-
-    def get_clip_range(self):
-        """
-            Get the current clip range.
-
-            Returns:
-                A tuple containing the clip range as (min, max).
-        """
-        min_value = self.clip_range[0](
-            iteration = self.status_dict["iteration"],
-            timestep  = self.status_dict["timesteps"])
-        max_value = self.clip_range[1](
-            iteration = self.status_dict["iteration"],
-            timestep  = self.status_dict["timesteps"])
-
-        return (min_value, max_value)
-
-    def _clip(self, val):
-        """
-            Perform the clip.
-
-            Arguments:
-                val    The value to be clipped.
-        """
-        raise NotImplementedError
-
-
-class ObservationClipper(GenericClipper):
-    """
-        An environment wrapper that clips observations.
-    """
-
-    def __init__(self,
-                 env,
-                 status_dict = {},
-                 clip_range  = (-10., 10.),
-                 **kw_args):
-        """
-            Initialize the wrapper.
-
-            Arguments:
-                env         The environment to wrap.
-                status_dict The training status dictionary. This is used when
-                            our clip range contains callables.
-                clip_range  The range to clip our observations into.
-        """
-        super(ObservationClipper, self).__init__(
-            env,
-            status_dict = status_dict,
-            clip_range  = clip_range,
-            **kw_args)
-
-    def step(self, action):
-        """
-            Take a single step in the environment using the given
-            action, and clip the resulting observation.
-
-            Arguments:
-                action    The action to take.
-
-            Returns:
-                The resulting observation, reward, done, and info tuple.
-        """
-        obs, reward, done, info = self.env.step(action)
-
-        obs = self._clip(obs)
-
-        return obs, reward, done, info
-
-    def reset(self):
-        """
-            Reset the environment, and clip the observation.
-
-            Returns:
-                The resulting observation.
-        """
-        obs = self.env.reset()
-
-        #
-        # In our multi-agent case, the global state is returned along
-        # with the local observation.
-        #
-        is_multi_agent = False
-        if type(obs) == type(tuple()):
-            is_multi_agent = True
-            obs, global_state = obs
-
-        obs = self._clip(obs)
-
-        if is_multi_agent:
-            return (obs, global_state)
-        return obs
-
-    def _clip(self, obs):
-        """
-            Perform the observation clip.
-
-            Arguments:
-                obs    The observation to clip.
-
-            Returns:
-                The clipped observation.
-        """
-        min_value, max_value = self.get_clip_range()
-        return np.clip(obs, min_value, max_value)
-
-
-class RewardClipper(GenericClipper):
-    """
-        A wrapper for clipping rewards.
-    """
-
-    def __init__(self,
-                 env,
-                 status_dict = {},
-                 clip_range  = (-10., 10.),
-                 **kw_args):
-        """
-            Initialize the wrapper.
-
-            Arguments:
-                env         The environment to wrap.
-                status_dict The training status dictionary. This is used when
-                            our clip range contains callables.
-                clip_range  The range to clip our rewards into.
-        """
-
-        super(RewardClipper, self).__init__(
-            env,
-            status_dict = status_dict,
-            clip_range  = clip_range,
-            **kw_args)
-
-    def step(self, action):
-        """
-            Take a single step in the environment using the given
-            action, and clip the resulting reward.
-
-            Arguments:
-                action    The action to take.
-
-            Returns:
-                The resulting observation, reward, done, and info tuple.
-        """
-        obs, reward, done, info = self.env.step(action)
-
-        if type(reward) == np.ndarray:
-            batch_size = reward.shape[0]
-
-            for r_idx in range(batch_size):
-                if "natural reward" not in info[r_idx]:
-                    info[r_idx]["natural reward"] = reward[r_idx]
-        else:
-            if "natural reward" not in info:
-                info["natural reward"] = reward
-
-        reward = self._clip(reward)
-
-        return obs, reward, done, info
-
-    def _clip(self, reward):
-        """
-            Perform the reward clip.
-
-            Arguments:
-                reward    The reward to clip.
-
-            Returns:
-                The clipped reward.
-        """
-        min_value, max_value = self.get_clip_range()
-        return np.clip(reward, min_value, max_value)
-
-
-class AugmentingEnvWrapper(IdentityWrapper):
-    """
-        This wrapper expects the environment to have a method named
-        'augment_observation' that can be utilized to augment an observation
-        to create a batch of obserations. Each instance of observation in the
-        batch will be coupled with an identical done and reward. This is to
-        help a policy learn that a particular augmentation does not affect
-        the learned policy.
-    """
-
-    def __init__(self,
-                 env,
-                 **kw_args):
-        """
-            Initialize the wrapper.
-
-            Arguments:
-                env   The environment to wrap.
-        """
-
-        super(AugmentingEnvWrapper, self).__init__(
-            env,
-            **kw_args)
-
-        self.test_idx = -1
-        aug_func      = getattr(env, "augment_observation", None)
-
-        if type(aug_func) == type(None):
-            msg  = "ERROR: env must define 'augment_observation' in order "
-            msg += "to qualify for the AugmentingEnvWrapper class."
-            rank_print(msg)
-            comm.Abort()
-
-        self.batch_size = self.aug_reset().shape[0]
-
-    def step(self, action):
-        """
-            Take a single step in the environment using the given
-            action. If we're in test mode, we don't augment. Otherwise,
-            call the aug_step method.
-
-            NOTE: the action is expected to be a SINGLE action. This does
-            not currently support multiple environment instances.
-
-            Arguments:
-                action    The action to take.
-
-            Returns:
-                The resulting observation(s), reward(s), done(s), and info(s).
-        """
-        if self.test_mode:
-            return self.aug_test_step(action)
-        return self.aug_step(action)
-
-    def aug_step(self, action):
-        """
-            Take a single step in the environment using the given
-            action, allow the environment to augment the returned
-            observation, and set up the return values as a batch.
-
-            NOTE: the action is expected to be a SINGLE action. This does
-            not currently support multiple environment instances.
-
-            Arguments:
-                action    The action to take.
-
-            Returns:
-                A batch of observations, rewards, and dones along
-                with a single info dictionary. The observations will
-                contain augmentations of the original observation.
-        """
-        obs, reward, done, info = self.env.step(action)
-
-        batch_obs  = self.env.augment_observation(obs)
-        batch_size = batch_obs.shape[0]
-
-        if "terminal observation" in info[0]:
-            batch_infos = np.array([None] * batch_size,
-                dtype=object)
-
-            terminal_obs = info[0]["terminal observation"]
-            terminal_obs = self.env.augment_observation(terminal_obs)
-
-            for i in range(batch_size):
-                i_info = info[0].copy()
-                i_info["terminal observation"] = terminal_obs[i].copy()
-                batch_infos[i] = i_info.copy()
-        else:
-            batch_infos = np.tile((info[0],), batch_size)
-
-        batch_rewards = np.tile((reward,), batch_size)
-        batch_dones   = np.tile((done,), batch_size).astype(bool)
-
-        batch_rewards = batch_rewards.reshape((batch_size, 1))
-        batch_dones   = batch_dones.reshape((batch_size, 1))
-        batch_infos   = batch_infos.reshape((batch_size))
-
-        return batch_obs, batch_rewards, batch_dones, batch_infos
-
-    def aug_test_step(self, action):
-        """
-            Take a single step in the environment using the given
-            action, allow the environment to augment the returned
-            observation. Since we're in test mode, we return a single
-            instance from the batch.
-
-            NOTE: the action is expected to be a SINGLE action. This does
-            not currently support multiple environment instances.
-
-            Arguments:
-                action    The action to take.
-
-            Returns:
-                Observation, reward, done, and info (possibly augmented).
-        """
-        obs, reward, done, info = self.env.step(action)
-
-        batch_obs  = self.env.augment_observation(obs)
-        batch_size = batch_obs.shape[0]
-
-        if self.test_idx < 0:
-            self.test_idx = np.random.randint(0, batch_size)
-
-        if "terminal observation" in info:
-            terminal_obs = info["terminal observation"]
-            terminal_obs = self.env.augment_observation(terminal_obs)
-            info["terminal observation"] = terminal_obs[self.test_idx].copy()
-
-        return batch_obs[self.test_idx], reward, done, info
-
-    def reset(self):
-        """
-            Reset the environment. If we're in test mode, we don't augment the
-            resulting observations. Otherwise, augment them before returning.
-
-            Returns:
-                The resulting observation(s).
-        """
-        if self.test_mode:
-            return self.aug_test_reset()
-        return self.aug_reset()
-
-    def aug_reset(self):
-        """
-            Reset the environment, and return a batch of augmented observations.
-
-            Returns:
-                The resulting observations.
-        """
-        obs = self.env.reset()
-
-        aug_obs_batch = self.env.augment_observation(obs)
-
-        return aug_obs_batch
-
-    def aug_test_reset(self):
-        """
-            Reset the environment, and return a single observations which may
-            or may not be augmented.
-
-            Returns:
-                The resulting observation (possibly augmented).
-        """
-        obs = self.env.reset()
-
-        aug_obs_batch = self.env.augment_observation(obs)
-        batch_size    = aug_obs_batch.shape[0]
-
-        if self.test_idx < 0:
-            self.test_idx = np.random.randint(0, batch_size)
-
-        return aug_obs_batch[self.test_idx]
-
-    def get_batch_size(self):
-        """
-            If any wrapped classes define this method, try to get the batch size
-            from them. Otherwise, assume we have a single environment.
-
-            Returns:
-                Return our batch size.
-        """
-        return self.batch_size
 
     def supports_batched_environments(self):
         """
